moving mapping functions to own files
