+++ mapper/mapper.go
@@ -6,20 +6,15 @@ import (
 	"net/url"
 	"regexp"
 	"sort"
-	"strconv"
 	"strings"
 
-	"github.com/ONSdigital/dp-api-clients-go/v2/dataset"
-	"github.com/ONSdigital/dp-api-clients-go/v2/filter"
 	"github.com/ONSdigital/dp-api-clients-go/v2/population"
 	"github.com/ONSdigital/dp-api-clients-go/v2/zebedee"
 	"github.com/ONSdigital/dp-cookies/cookies"
 	"github.com/ONSdigital/dp-frontend-filter-flex-dataset/helpers"
 	"github.com/ONSdigital/dp-frontend-filter-flex-dataset/model"
-	"github.com/ONSdigital/dp-frontend-filter-flex-dataset/pagination"
 	"github.com/ONSdigital/dp-renderer/helper"
 	coreModel "github.com/ONSdigital/dp-renderer/model"
-	"github.com/ONSdigital/log.go/v2/log"
 )
 
 // Constants...
@@ -37,374 +32,6 @@ const (
 	areaPageType          = "area_type_options"
 	reviewPageType        = "review_changes"
 )
-
-// CreateFilterFlexOverview maps data to the Overview model
-func CreateFilterFlexOverview(req *http.Request, basePage coreModel.Page, lang, path string, queryStrValues []string, filterJob filter.GetFilterResponse, filterDims []model.FilterDimension, datasetDims dataset.VersionDimensions, hasNoAreaOptions, isMultivariate bool, eb zebedee.EmergencyBanner, serviceMsg string) model.Overview {
-	p := model.Overview{
-		Page: basePage,
-	}
-	mapCommonProps(req, &p.Page, reviewPageType, "Review changes", lang, serviceMsg, eb)
-	p.FilterID = filterJob.FilterID
-	dataset := filterJob.Dataset
-	p.IsMultivariate = isMultivariate
-
-	p.Breadcrumb = []coreModel.TaxonomyNode{
-		{
-			Title: helper.Localise("Back", lang, 1),
-			URI: fmt.Sprintf("/datasets/%s/editions/%s/versions/%s",
-				dataset.DatasetID,
-				dataset.Edition,
-				strconv.Itoa(dataset.Version)),
-		},
-	}
-
-	for _, dim := range filterDims {
-		pageDim := model.Dimension{}
-		pageDim.Name = cleanDimensionLabel(dim.Label)
-		pageDim.IsAreaType = *dim.IsAreaType
-		pageDim.OptionsCount = dim.OptionsCount
-		pageDim.ID = dim.ID
-		pageDim.URI = fmt.Sprintf("%s/%s", path, dim.Name)
-		pageDim.IsChangeCategories = isMultivariate
-		q := url.Values{}
-		midFloor, midCeiling := getTruncationMidRange(dim.OptionsCount)
-
-		var displayedOptions []string
-		if len(dim.Options) > 9 && !helpers.HasStringInSlice(dim.Name, queryStrValues) && !*dim.IsAreaType {
-			displayedOptions = append(displayedOptions, dim.Options[:3]...)
-			displayedOptions = append(displayedOptions, dim.Options[midFloor:midCeiling]...)
-			displayedOptions = append(displayedOptions, dim.Options[len(dim.Options)-3:]...)
-			q.Add(queryStrKey, dim.Name)
-			helpers.PersistExistingParams(queryStrValues, queryStrKey, dim.Name, q)
-			pageDim.IsTruncated = true
-		} else {
-			helpers.PersistExistingParams(queryStrValues, queryStrKey, dim.Name, q)
-			displayedOptions = dim.Options
-			pageDim.IsTruncated = false
-		}
-
-		pageDim.Options = append(pageDim.Options, displayedOptions...)
-		pageDim.TruncateLink = generateTruncatePath(path, dim.ID, q)
-		p.Dimensions = append(p.Dimensions, pageDim)
-	}
-
-	sort.Slice(p.Dimensions, func(i, j int) bool {
-		return p.Dimensions[i].IsAreaType
-	})
-
-	coverage := []model.Dimension{
-		{
-			IsCoverage:        true,
-			IsDefaultCoverage: hasNoAreaOptions,
-			URI:               fmt.Sprintf("%s/geography/coverage", path),
-			Options:           p.Dimensions[0].Options,
-			ID:                "coverage",
-		},
-	}
-	temp := append(coverage, p.Dimensions[1:]...)
-	p.Dimensions = append(p.Dimensions[:1], temp...)
-
-	// TODO: Temporarily removing mapping as new endpoints are required to return reliable dataset information
-
-	return p
-}
-
-// CreateCategorisationsSelector maps data to the Selector model
-func CreateCategorisationsSelector(req *http.Request, basePage coreModel.Page, dimLabel, lang, filterID, dimId, serviceMsg string, eb zebedee.EmergencyBanner, cats population.GetCategorisationsResponse, isValidationError bool) model.Selector {
-	p := model.Selector{
-		Page: basePage,
-	}
-	mapCommonProps(req, &p.Page, "filter-flex-selector", cleanDimensionLabel(dimLabel), lang, serviceMsg, eb)
-	p.Breadcrumb = []coreModel.TaxonomyNode{
-		{
-			Title: helper.Localise("Back", lang, 1),
-			URI:   fmt.Sprintf("/filters/%s/dimensions", filterID),
-		},
-	}
-	p.LeadText = helper.Localise("SelectCategoriesLeadText", lang, 1)
-	p.InitialSelection = dimId
-
-	var selections []model.Selection
-	for _, cat := range cats.Items {
-		cats := []string{}
-		for _, c := range sortCategoriesByID(cat.Categories) {
-			cats = append(cats, c.Label)
-		}
-		selections = append(selections, mapCats(cats, req.URL.Query()["showAll"], lang, req.URL.Path, cat.ID))
-	}
-	p.Selections = selections
-
-	if isValidationError {
-		p.Page.Error = coreModel.Error{
-			Title: p.Page.Metadata.Title,
-			ErrorItems: []coreModel.ErrorItem{
-				{
-					Description: coreModel.Localisation{
-						LocaleKey: "SelectCategoriesError",
-						Plural:    1,
-					},
-					URL: "#categories-error",
-				},
-			},
-			Language: lang,
-		}
-		p.ErrorId = "categories-error"
-	}
-
-	return p
-}
-
-// CreateAreaTypeSelector maps data to the Selector model
-func CreateAreaTypeSelector(req *http.Request, basePage coreModel.Page, lang, filterID string, areaType []population.AreaType, fDim filter.Dimension, lowest_geography, releaseDate string, dataset dataset.DatasetDetails, isValidationError, hasOpts bool, serviceMsg string, eb zebedee.EmergencyBanner) model.Selector {
-	p := model.Selector{
-		Page: basePage,
-	}
-	mapCommonProps(req, &p.Page, areaPageType, areaTypeTitle, lang, serviceMsg, eb)
-	p.Breadcrumb = []coreModel.TaxonomyNode{
-		{
-			Title: helper.Localise("Back", lang, 1),
-			URI:   fmt.Sprintf("/filters/%s/dimensions", filterID),
-		},
-	}
-	p.LeadText = helper.Localise("SelectAreaTypeLeadText", lang, 1)
-	if hasOpts {
-		p.Panel = mapPanel(coreModel.Localisation{
-			LocaleKey: "ChangeAreaTypeWarning",
-			Plural:    1,
-		}, lang, []string{"ons-u-mb-l"})
-	}
-
-	if isValidationError {
-		p.Page.Error = coreModel.Error{
-			Title: p.Page.Metadata.Title,
-			ErrorItems: []coreModel.ErrorItem{
-				{
-					Description: coreModel.Localisation{
-						LocaleKey: "SelectAreaTypeError",
-						Plural:    1,
-					},
-					URL: "#area-type-error",
-				},
-			},
-			Language: lang,
-		}
-		p.ErrorId = "area-type-error"
-	}
-
-	selections := mapAreaTypesToSelection(sortAreaTypes(areaType))
-
-	if lowest_geography != "" {
-		var filtered_selections []model.Selection
-		var lowest_found = false
-		for _, selection := range selections {
-			if !lowest_found {
-				filtered_selections = append(filtered_selections, selection)
-				lowest_found = selection.Value == lowest_geography
-			}
-		}
-		p.Selections = filtered_selections
-	} else {
-		p.Selections = selections
-	}
-
-	p.InitialSelection = fDim.ID
-	p.IsAreaType = true
-
-	p.DatasetId = dataset.ID
-	p.DatasetTitle = dataset.Title
-	p.ReleaseDate = releaseDate
-
-	return p
-}
-
-// CreateGetCoverage maps data to the coverage model
-func CreateGetCoverage(req *http.Request, basePage coreModel.Page, lang, filterID, geogName, nameQ, parentQ, parentArea, setParent, coverage, dim, geogID, releaseDate, serviceMsg string, eb zebedee.EmergencyBanner, dataset dataset.DatasetDetails, areas population.GetAreasResponse, opts []model.SelectableElement, parents population.GetAreaTypeParentsResponse, hasFilterByParent, hasValidationErr bool, currentPage int) model.Coverage {
-	p := model.Coverage{
-		Page: basePage,
-	}
-	mapCommonProps(req, &p.Page, coveragePageType, coverageTitle, lang, serviceMsg, eb)
-	p.Breadcrumb = []coreModel.TaxonomyNode{
-		{
-			Title: helper.Localise("Back", lang, 1),
-			URI:   fmt.Sprintf("/filters/%s/dimensions", filterID),
-		},
-	}
-	geography := helpers.Pluralise(req, geogName, lang, areaTypePrefix, pluralInt)
-	if geography == "" {
-		log.Info(req.Context(), "pluralisation lookup failed, reverting to initial input", log.Data{
-			"initial_input": geogName,
-		})
-		geography = geogName
-	}
-
-	p.Geography = strings.ToLower(geography)
-	p.CoverageType = coverage
-	p.Dimension = dim
-	p.GeographyID = geogID
-	p.SetParent = setParent
-	p.NameSearch = model.SearchField{
-		Name:     nameSearchFieldName,
-		ID:       nameSearch,
-		Value:    nameQ,
-		Language: lang,
-		Label:    helper.Localise("CoverageSearchLabel", lang, 1),
-	}
-	p.ParentSearch = model.SearchField{
-		Name:     parentSearchFieldName,
-		ID:       parentSearch,
-		Value:    parentQ,
-		Language: lang,
-		Label:    helper.Localise("CoverageSearchLabel", lang, 1),
-	}
-
-	p.DatasetId = dataset.ID
-	p.DatasetTitle = dataset.Title
-	p.ReleaseDate = releaseDate
-
-	if len(parents.AreaTypes) > 1 && parentArea == "" {
-		p.ParentSelect = []model.SelectableElement{
-			{
-				Text:       helper.Localise("CoverageSelectDefault", lang, 1),
-				IsSelected: true,
-				IsDisabled: true,
-			},
-		}
-	}
-	for _, parent := range parents.AreaTypes {
-		var sel model.SelectableElement
-		sel.Text = parent.Label
-		sel.Value = parent.ID
-		if parentArea == parent.ID {
-			sel.IsSelected = true
-		}
-		p.ParentSelect = append(p.ParentSelect, sel)
-	}
-
-	var isParentSearch bool
-	if coverage == parentSearch {
-		isParentSearch = true
-	}
-	var results []model.SelectableElement
-	for _, area := range areas.Areas {
-		var result model.SelectableElement
-		result.Text = area.Label
-		result.Value = area.ID
-		result.Name = getAddOptionStr(isParentSearch)
-		for _, opt := range opts {
-			if opt.Value == area.ID {
-				result.IsSelected = true
-				result.Name = "delete-option"
-				break
-			}
-		}
-		results = append(results, result)
-	}
-
-	totalPages := pagination.GetTotalPages(areas.TotalCount, areas.Limit)
-	var paginatedResults coreModel.Pagination
-	if totalPages > 1 {
-		paginatedResults = coreModel.Pagination{
-			CurrentPage:       currentPage,
-			TotalPages:        totalPages,
-			Limit:             areas.Limit,
-			FirstAndLastPages: pagination.GetFirstAndLastPages(req, totalPages),
-			PagesToDisplay:    pagination.GetPagesToDisplay(currentPage, totalPages, req),
-		}
-	}
-
-	if len(opts) > 0 && hasFilterByParent {
-		p.CoverageType = parentSearch
-		p.ParentSearchOutput.Selections = opts
-		p.ParentSearchOutput.SelectionsTitle = helper.Localise("AreasAddedTitle", lang, len(opts))
-		p.OptionType = parentSearch
-	} else if len(opts) > 0 {
-		p.CoverageType = nameSearch
-		p.NameSearchOutput.Selections = opts
-		p.ParentSearchOutput.SelectionsTitle = helper.Localise("AreasAddedTitle", lang, len(opts))
-		p.OptionType = nameSearch
-	}
-
-	switch coverage {
-	case nameSearch:
-		p.CoverageType = nameSearch
-		p.NameSearchOutput.Results = results
-		p.NameSearchOutput.HasNoResults = len(p.NameSearchOutput.Results) == 0
-		p.NameSearchOutput.Pagination = paginatedResults
-	case parentSearch:
-		p.CoverageType = parentSearch
-		p.ParentSearchOutput.Results = results
-		p.ParentSearchOutput.HasNoResults = len(p.ParentSearchOutput.Results) == 0 && !hasValidationErr
-		p.ParentSearchOutput.Pagination = paginatedResults
-	}
-
-	if hasValidationErr {
-		p.Page.Error = coreModel.Error{
-			Title: p.Metadata.Title,
-			ErrorItems: []coreModel.ErrorItem{
-				{
-					Description: coreModel.Localisation{
-						LocaleKey: "CoverageSelectDefault",
-						Plural:    1,
-					},
-					URL: "#coverage-error",
-				},
-			},
-			Language: lang,
-		}
-	}
-
-	p.IsSelectParents = len(parents.AreaTypes) > 0
-
-	return p
-}
-
-// CreateGetChangeDimensions maps data to the ChangeDimensions model
-func CreateGetChangeDimensions(req *http.Request, basePage coreModel.Page, lang, fid, q, formAction, serviceMsg string, eb zebedee.EmergencyBanner, dims []model.FilterDimension, pDims, results population.GetDimensionsResponse) model.ChangeDimensions {
-	p := model.ChangeDimensions{
-		Page: basePage,
-	}
-	p.Breadcrumb = []coreModel.TaxonomyNode{
-		{
-			Title: helper.Localise("Back", lang, 1),
-			URI:   fmt.Sprintf("/filters/%s/dimensions", fid),
-		},
-	}
-	mapCommonProps(req, &p.Page, "change_variables", "Add or remove variables", lang, serviceMsg, eb)
-	p.Panel = mapPanel(coreModel.Localisation{
-		LocaleKey: "DimensionsChangeWarning",
-		Plural:    1,
-	}, lang, []string{"ons-u-mb-s"})
-	p.FormAction = formAction
-
-	selections := []model.SelectableElement{}
-	for _, dim := range dims {
-		if !*dim.IsAreaType {
-			selections = append(selections, model.SelectableElement{
-				Text:  cleanDimensionLabel(dim.Label),
-				Value: dim.ID,
-				Name:  "delete-option",
-			})
-		}
-	}
-	p.Output.Selections = selections
-	p.Output.SelectionsTitle = "Variables added"
-	p.Search = model.SearchField{
-		Name:     "q",
-		ID:       "dimensions-search",
-		Language: lang,
-		Value:    q,
-		Label:    helper.Localise("DimensionsSearchLabel", lang, 1),
-	}
-
-	browseResults := mapDimensionsResponse(pDims, &selections)
-	searchResults := mapDimensionsResponse(results, &selections)
-
-	p.Output.Results = browseResults
-	p.SearchOutput.Results = searchResults
-	p.SearchOutput.HasNoResults = len(p.SearchOutput.Results) == 0 && formAction == "search"
-
-	return p
-}
 
 // mapDimensionsResponse returns a sorted array of selectable elements
 func mapDimensionsResponse(pDims population.GetDimensionsResponse, selections *[]model.SelectableElement) []model.SelectableElement {
